return boolean for A*
